backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
python-multipart==0.0.9

# ── RAG / Embeddings ─────────────────────────────────────────
faiss-cpu>=1.7.4
sentence-transformers==2.7.0
pypdf==4.2.0

# ── LLM ──────────────────────────────────────────────────────
google-genai

# ── Language ─────────────────────────────────────────────────
langdetect==1.0.9
deep-translator==1.11.4

# ── HTTP / APIs ──────────────────────────────────────────────
httpx==0.27.0
requests==2.31.0

# ── Auth & Database ──────────────────────────────────────────
sqlalchemy>=2.0.49
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# ── Utils ────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
tenacity>=8.0.0
